requirements.txt
flask
PyYAML
waitress
Flask-HTTPAuth
requests